Как в демонстрационном проекте

+++ app/src/main/java/ru/netology/statsview/utils/AndroidUtils.kt
@@ -4,6 +4,6 @@ import android.content.Context
 import kotlin.math.ceil
 
 object AndroidUtils {
-    fun dp(context: Context, dp: Int): Int =
+    fun dp(context: Context, dp: Float): Int =
         ceil(context.resources.displayMetrics.density * dp).toInt()
 }
